+++ monorepo-builder.yml
@@ -139,6 +139,7 @@ parameters:
     src/Unsplash: 'git@github.com:SocialiteProviders/Unsplash.git'
     src/Untappd: 'git@github.com:SocialiteProviders/Untappd.git'
     src/Venmo: 'git@github.com:SocialiteProviders/Venmo.git'
+    src/Vercel: 'git@github.com:SocialiteProviders/Vercel.git'
     src/VersionOne: 'git@github.com:SocialiteProviders/VersionOne.git'
     src/Vimeo: 'git@github.com:SocialiteProviders/Vimeo.git'
     src/VKontakte: 'git@github.com:SocialiteProviders/VKontakte.git'
@@ -158,5 +159,4 @@ parameters:
     src/Yiban: 'git@github.com:SocialiteProviders/Yiban.git'
     src/YouTube: 'git@github.com:SocialiteProviders/YouTube.git'
     src/Zendesk: 'git@github.com:SocialiteProviders/Zendesk.git'
-    src/Zeit: 'git@github.com:SocialiteProviders/Zeit.git'
     src/Zoho: 'git@github.com:SocialiteProviders/Zoho.git'
